Rename and move class
